chore: added `Dismissable` component
